update: merge code SetOp.kt and add mobGriefing options
